Adding in more Account Creation Parameters. Email must have a real email account.
Wanted to try to see if I could make it so that the user must input a legitimate email account. This would include Top Level Domain tags (.com, .net, etc.) and also place in a @ symbol. I will more than likely be importing the email verifier function as a replacement     : l